Code refactoring and normalisation fix
The energy histogram is not normalised in the time dimension.

Robert's expalation:  I think all you need to do is remove
scale = time_bin_width / nu.year  # Convert from [per year] to [per time_bin_width]
Whilst in principle keeping this and removing  / self.n_time_bins would solve things at the level of source.mu_before_efficiencies(), it will mess up the tensor-based calculation of the differential rate, as the thing that is queried should be the spectrum corresponding to a given event time (such that the sum of the spectrum gives the expected counts within that energy range, after scaling by the exposure)
